refactor: 명료한 네이밍으로 수정
