+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as Footer } from './Footer';
 export { default as Header } from './Header';
+export { default as MovieSection } from './MovieSection';
 export { default as Movies } from './Movies';
-export { default as ThumbnailList } from './ThumbnailList';
 export { default as MovieDetailModal } from './MovieDetailModal';
 export { default as IntersectionObserble } from './IntersectionObserble';
 export { default as Toast } from './Toast';
